서버 및 네트워크 관련 클래스 수정: SessionGroup 클래스 추가, Session 클래스의 Send 및 DoSend 메서드 개선, Server 클래스에 세션 관리 메서드 추가, Job 시스템의 Execute 메서드 수정, CMake 설정 업데이트

+++ src/lib-core/include/core/Types.hpp
@@ -30,7 +30,7 @@ using SocketFD = int;
 using SocketFD = SOCKET;
 #endif
 
-using SessionID = Int32;
+using SessionID = Int64;
 
 const SocketFD sINVALID_SOCKET_FD = -1;
 const Int32 sINVALID_ID = -1;
